requirements.txt
discord.py==1.4.1
aiohttp
websockets
chardet